Producing solutions for simple problems
case NOT : changed to "run() && !(run())" from "!(run())"
because it becomes a binary operator this way.

found out that randomnumber is the quantity of terminals rather than variables, it needs some terminals to solve problem.txt but can solve "or" or "triple and" without.

fitness starts at 1 and is multiplied by 2 every time the program matches a truth table. Therefore if there's 16 fitness cases the max fitness is 2 ^ 16.